pyproject.toml
[build-system]
requires = ["scikit-build-core>=0.11.6", "setuptools>=73.0.1"]
build-backend = "scikit_build_core.setuptools.build_meta"

[tool.scikit-build]
wheel.packages = ["src/fandango"]

# If you update this info, please also update the version in setup.py
[project]
name = "fandango-fuzzer"
version = "1.1.0-rc.1"
authors = [
    { name = "José Antonio Zamudio Amaya", email = "jose.zamudio@cispa.de" },
    { name = "Marius Smytzek", email = "marius.smytzek@cispa.de" },
    { name = "Andreas Zeller", email = "andreas.zeller@cispa.de" },
]
description = "Fandango produces myriads of high-quality random inputs to test programs, giving users unprecedented control over format and shape of the inputs."
readme = "README.md"
license = { file = "LICENSE.md" }
requires-python = ">=3.10,<4.0"
classifiers = [
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "ansi_styles>=0.2.2",
    "antlr4-python3-runtime>=4.13",
    "cachedir_tag>=0.0.3",
    "dill>=0.3.7",
    "exrex>=0.12.0",
    "gnureadline>=8.2.13 ; platform_system != 'Windows' and implementation_name != 'pypy'",
    "lxml>=5.3.2",
    "py010parser>=0.1.18",
    "pyreadline3>=3.5.4; platform_system == 'Windows'",
    "regex>=2024.11.6",
    "tdigest>=0.5.2.2",
    "thefuzz>=0.22",
    "xdg-base-dirs>=6.0.2",
    "beartype>=0.21.0,!=0.22.6,!=0.22.7",
]

[project.optional-dependencies]
test = [
    "Faker>=30.4.0",
    "pytest>=8.3.5",
    "pytest-benchmark>=5.1.0",
    "pytest-cov>=4.1.0",
    "pytest-html>=3.2.0",
    "pytest-pythonhashseed>=2.0.0",
    "pytest-rerunfailures>=11.1.2",
    "pytest-timeout>=2.4.0",
    "pytest-xdist>=3.7.0",
    "python-dateutil>=2.9.0.post0",
    "docutils>=0.20.1",
    "tccbox>=2025.6.6",             # We have upgraded from ==2024.7.1 (the version ISLa used) because it is not compatible with Python 3.12 anymore :(
]

development = [
    "fandango-fuzzer[test]",
    "black>=25.1.0",
    "mypy>=1.16.0",
    "pre-commit>=4.2.0",
    "pygls>=2.0.0a4",           # language server
    "speedy-antlr-tool>=1.4.3",
    "types-setuptools>=73.0.1", # enable mypy to check setup.py
    "scikit-build-core>=0.11.6", # For type checking
]

book = [
    "fandango-fuzzer[development]", # need black and faker
    "aiosmtpd>=1.4.6",              # example server for protocol testing
    "docutils>=0.20.1",             # Jupyter-book needs docutils<=0.20
    "ghp-import>=2.1.0",
    "graphviz>=0.20.3",
    "jupyter-book>=1.0.4,<2.0.0",
    "pagelabels>=1.2.1",
    "pyppeteer>=2.0.0",
    "sphinxcontrib-mermaid>=1.0.0",
]

[project.urls]
homepage = "https://fandango-fuzzer.github.io/"
repository = "https://github.com/fandango-fuzzer/fandango/"
"Bug Tracker" = "https://github.com/fandango-fuzzer/fandango/issues"

[project.scripts]
fandango = "fandango.cli:main"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
follow_untyped_imports = true
check_untyped_defs = true
enable_error_code = ['ignore-without-code']
# See https://mypy.readthedocs.io/en/stable/config_file.html#confval-exclude
exclude = [
    'src/fandango/language/parser/FandangoLexer\.py$',
    'src/fandango/language/parser/FandangoParser\.py$',
    'src/fandango/language/parser/FandangoParserVisitor\.py$',
    'src/fandango/language/parser/sa_fandango\.py$',
    'src/fandango/converters/antlr/ANTLRv4[^.]*\.py$',
    'src/fandango/converters/antlr/LexerAdaptor[^.]*\.py$',
]

[tool.pytest.ini_options]
addopts = "-n auto --timeout 300 --durations=10 --pythonhashseed=1 --benchmark-max-time=10 --benchmark-disable"
filterwarnings = [
    "error",                                                  # warnings are errors
    "ignore::pytest_benchmark.logger.PytestBenchmarkWarning", # benchmarks don't run with xdist, that's fine, they are run manually in the CI
    "ignore::ResourceWarning",                                # windows keeps a socket open for some reason
]
